move urlOpts
