+++ packages/expo-cli/src/commands/start/startAsync.ts
@@ -6,11 +6,11 @@ import { Project, UnifiedAnalytics, UrlUtils, Versions } from 'xdl';
 
 import getDevClientProperties from '../../analytics/getDevClientProperties';
 import Log from '../../log';
-import urlOpts from '../../urlOpts';
 import { assertProjectHasExpoExtensionFilesAsync } from '../utils/deprecatedExtensionWarnings';
 import { profileMethod } from '../utils/profileMethod';
 import * as sendTo from '../utils/sendTo';
 import { ensureTypeScriptSetupAsync } from '../utils/typescript/ensureTypeScriptSetup';
+import urlOpts from '../utils/urlOpts';
 import { validateDependenciesVersionsAsync } from '../utils/validateDependenciesVersions';
 import * as TerminalUI from './TerminalUI';
 import { installCustomExitHook, installExitHooks } from './installExitHooks';
