+++ packages/expo-cli/src/commands/url/urlAsync.ts
@@ -4,7 +4,7 @@ import { ProjectSettings, UrlUtils } from 'xdl';
 
 import CommandError from '../../CommandError';
 import Log from '../../log';
-import urlOpts, { URLOptions } from '../../urlOpts';
+import urlOpts, { URLOptions } from '../utils/urlOpts';
 import printRunInstructionsAsync from './printRunInstructionsAsync';
 
 type ProjectUrlOptions = Command & {
