+++ packages/expo-cli/src/commands/utils/urlOpts.ts
@@ -9,9 +9,9 @@ import {
   Webpack,
 } from 'xdl';
 
-import CommandError, { AbortCommandError } from './CommandError';
-import Log from './log';
-import { getOptionalDevClientSchemeAsync } from './utils/schemes';
+import CommandError, { AbortCommandError } from '../../CommandError';
+import Log from '../../log';
+import { getOptionalDevClientSchemeAsync } from '../../utils/schemes';
 
 // NOTE: if you update this, you should also update assertValidOptions in UrlUtils.ts
 export type URLOptions = {
