+++ packages/expo-cli/src/exp.ts
@@ -36,8 +36,8 @@ import StatusEventEmitter from './analytics/StatusEventEmitter';
 import { registerCommands } from './commands';
 import { learnMore } from './commands/utils/TerminalLink';
 import { profileMethod } from './commands/utils/profileMethod';
+import urlOpts from './commands/utils/urlOpts';
 import Log from './log';
-import urlOpts from './urlOpts';
 import { matchFileNameOrURLFromStackTrace } from './utils/matchFileNameOrURLFromStackTrace';
 import { logNewSection, ora } from './utils/ora';
 import update from './utils/update';
